0005-DNS: Cache names longer with confidence
Web server and DNS may be hijacked and return the wrong Dat key. This new
caching method allows for name resolution caching to grow linearly as
confidence in the returned result increases over time.

E.g. if a name was first resolved three days ago, it can be cached for up
to six additional days if we're confident in the resolver result. After
two weeks, the name can be cached for four weeks  and so on. Domain owner
still controls maximum caching time with TTL.

Increases performance and privacy while making clients less susceptible to
centralized/decentralized service disruptions (such as a DDoS or a server
running out of memory.) Reduces dependency on servers.